fix: upcoming races and distance in correct format

+++ src/main/java/ucll/be/procyclingscraper/model/Stage.java
@@ -24,9 +24,9 @@ public class Stage {
     private String arrival;
     private String date;
     private String startTime;
-    private Integer distance;
+    private Double distance;
     private String stageUrl;
-    private Integer verticalMeters;
+    private Double verticalMeters;
 
     @OneToMany(mappedBy = "stage", cascade = CascadeType.ALL)
     private List<StageResult> results = new ArrayList<>();
